initial split of the Expr file
