+++ scala-backend/app/saladbar/src/main/scala/Bop.scala
@@ -1,0 +1,77 @@
+package saladbar
+
+
+/**
+  * Bop
+  * 
+  * a collection of binary operations
+  */
+sealed trait Bop {
+    override def toString: String
+}
+
+
+case object And extends Bop{
+    override def toString: String = "&&"
+}
+
+
+case object Or extends Bop{
+    override def toString: String = "||"
+}
+
+
+case object Plus extends Bop {
+    override def toString: String = "+"
+}
+
+
+case object Times extends Bop{
+    override def toString: String = "*"
+}
+
+
+case object Minus extends Bop {
+    override def toString: String = "-"
+}
+
+
+case object Geq extends Bop {
+    override def toString: String = ">="
+}
+
+
+case object Gt extends Bop {
+    override def toString: String = ">"
+}
+
+
+case object Leq extends Bop {
+    override def toString: String = "<="
+}
+
+
+case object Lt extends Bop {
+    override def toString: String = "<"
+}
+
+
+case object Eq extends Bop {
+    override def toString: String = "=="
+}
+
+
+case object Eqq extends Bop {
+    override def toString: String = "==="
+}
+
+
+case object Neq extends Bop {
+    override def toString: String = "!="
+}
+
+
+case object Neqq extends Bop {
+    override def toString: String = "!=="
+}
+
