+++ scala-backend/app/saladbar/src/main/scala/ExprToStringError.scala
@@ -1,0 +1,16 @@
+package saladbar
+
+
+/**
+  * ExprToStringError
+  *
+  * @param msg
+  */
+case class ExprToStringError(msg: String) extends Exception {
+
+
+    override def toString(): String = s"EXPR TO STRING: $msg"
+
+    
+}
+
