+++ scala-backend/app/saladbar/src/main/scala/LetRec.scala
@@ -1,0 +1,48 @@
+package saladbar
+
+
+/**
+  * LetRec(id_functionName, closure, e2)
+  * from concrete syntax: letrec id_functionName = closure in e2
+  *
+  * @param id_functionName
+  * @param closure
+  * @param e2
+  */
+case class LetRec(id_functionName: String,
+        closure: Closure,
+        e2: Expr) extends Expr {
+
+
+    override def toString: String = s"letrec $id_functionName = $closure in $e2"
+
+
+    override def step[A](evalConditions: EvalConditions)(sc: Expr => A): A = {
+        val Closure(id_parameter, e_functionBody, env) = closure
+        val newEnv: Environment = ExtendRec(id_functionName, id_parameter, e_functionBody, env)
+        val v: Value = Closure(id_parameter, e_functionBody, newEnv)
+        e2.substitute(evalConditions, id_functionName, v)(sc)
+    }
+
+
+    override def substitute[A](evalConditions: EvalConditions, x: String, esub: Expr)(sc: Expr => A): A = {
+        // letrec f = cl(x) e {env} in e2
+        val Closure(id_parameter, e_functionBody, envcl) = closure
+        if (x == id_functionName) {
+            sc(this)
+        } else if (x == id_parameter) {
+            e2.substitute(evalConditions, x, esub){
+                e2p => sc(LetRec(id_functionName, closure, e2p))
+            }
+        } else {
+            e_functionBody.substitute(evalConditions, x, esub){
+                newFuncitonBody => e2.substitute(evalConditions, x, esub){
+                    e2p => sc(LetRec(id_functionName, Closure(id_parameter, newFuncitonBody, envcl), e2p))
+                }
+            }
+        }
+    }
+
+
+}
+
