+++ scala-backend/app/saladbar/src/main/scala/LookupError.scala
@@ -1,0 +1,12 @@
+package saladbar
+
+
+/**
+  * LookupError
+  *
+  * @param x
+  */
+case class LookupError(x: String) extends Exception {
+    override def toString(): String = s"Expected $x"
+}
+
